+++ src/Entity/UserRole.php
@@ -9,5 +9,6 @@ enum UserRole: string
     case SuperAdmin = 'ROLE_SUPER_ADMIN';
     case OverviewEditor = 'ROLE_OVERVIEW_EDITOR';
     case OverviewViewer = 'ROLE_OVERVIEW_VIEWER';
+    case OverviewSearcher = 'ROLE_OVERVIEW_SEARCHER';
     case ProcessStepRunner = 'ROLE_PROCESS_STEP_RUNNER';
 }